move editor field to form dir
